services/gallery-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
Pillow==10.1.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
prometheus-fastapi-instrumentator==6.1.0
